BLIP Model/requirements.txt
streamlit
torch
torchvision
transformers
Pillow